Add trace logging
